requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.0
astroid==4.0.4
attrs==26.1.0
awsebcli==3.27.1
bcrypt==5.0.0
billiard==4.2.4
black==25.12.0
blessed==1.27.0
boto3==1.42.84
botocore==1.42.84
celery==5.6.2
cement==2.10.14
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.13.4
cryptography==46.0.6
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.1
distro==1.9.0
Django==5.2.12
django-cors-headers==4.9.0
django-rest-swagger==2.2.0
django-storages==1.14.6
djangorestframework==3.16.1
fabric==3.2.2
factory_boy==3.3.3
Faker==40.11.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
isort==8.0.1
itypes==1.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
kombu==5.6.2
legacy-cgi==2.6.4
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
openai==2.30.0
openapi-codec==1.3.2
packaging==25.0
paramiko==4.0.0
pathspec==0.12.1
pillow==12.2.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.5
pylint-django==2.7.0
pylint-plugin-utils==0.9.0
PyNaCl==1.6.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.3.0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
ruff==0.15.6
s3transfer==0.16.0
semantic-version==2.10.0
setuptools==66.1.1
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
standardwebhooks==1.0.1
svix==1.89.0
termcolor==2.5.0
tomlkit==0.14.0
tqdm==4.67.3
types-Deprecated==1.3.1.20260130
types-python-dateutil==2.9.0.20260323
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.12.0
wrapt==2.1.2
